pyproject.toml
[tool.poetry]
name = "web-hw-14"
version = "0.1.0"
package-mode = false
description = "add new function"
authors = ["Vlad <vlad111121562@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.7.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "^5.0.3"
jinja2 = "^3.1.3"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.39.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
httpx = "^0.27.0"
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
